Initial implementation of page for report info
