+++ .env.example
@@ -6,6 +6,7 @@ VITE_API_HOST="${VITE_APP_DOMAIN}"
 VITE_ENABLE_USER_REGISTER=false
 
 VITE_LOCALE=en-US
+VITE_TIMEZONE=UTC
 
 VITE_HCAPTCHA_SITE_KEY=
 VITE_HCAPTCHA_SIZE=invisible
